fix(terraform/github): Cleanup terraform resource names
The GitHub repo name was being used directly for the terraform resource
name. This is not ideal because terraform does not allow for characters
such as '.' in resource names. This prevents the exact error from '.' by
converting it to '-'. Since it is still useful to know the actual
repository name, this has been added as a value to the config,
`terraformName`. This value is what should be used when interacting with
the terraform resource.