chore(nomad): give factorio more ram
Paperless has been disabled as it isn't being used. Factorio was being a
bit buggy and the server was crashing frequently. The process seemed
unable to write to its log. I've added more ram, removed the nomad
allocation, and restored it. All seems to be working well now. The
server hasn't crashed since making this change