chore: update dependencies
Updates for nixpkgs, home-manager, and hyprland. Configuration changes
have been made that enable steam to run on hyprland again. Gamescope has
been added with the primary monitor's resolution.

Patched nomad has been disabled. The patch will need to be updated to
work with the new version, or it should be removed.