chore: use proper flake-builder image
The image is now created as part of the upstream project. A nomad job
has been created to properly run the server. Resource amounts have been
set at a range that seems to allow for nix build to succeed. Builds can
be killed if they run out of memory. Nix is configured so that build
commands will work in the container.