build(terraform): switch to opentofu
Opentofu is used in place of terraform. Many options have been specified
explicitly for compatibility. This repository still depends on terraform
cloud so the url has been set to reflect this. All configurations have
been tested against opentofu and the terraform lockfiles reflect this
change.

The thoenix cli has been updated so it now defaults to using tofu.

Many of the names used in this repository remain to be `terraform`
because this change makes no attempts to change them. The only
difference currently is that the package is switched out to opentofu and
the configurations contained within work properly with it.