refactor(nixos/modules): update cloudhost/hetzner
This is in response to an error that was likely brought in from a
nixpkgs update. It appears that using `boot.initrd.postDeviceCommands`
is not possible in recent versions. Following the suggestion the zfs
rollback command is now a systemd service that is run.

This may or may not work properly. The configuration was never tested
for the ability of this section of code to work and so no impermanence
is claimed in this configuration. It was also noticed that the zfs pool
that is used is `footfs` instead of `rootfs`. This has not been
investigated fully. The update only does enough to make the
configurations build again.