fix(home-manager): disable manuals
sourcehut is currently down, which is causing home-manager builds to
fail. This results in manuals being unable to be retrieved for when
building a home-manager configuration.

This is a temporary fix and should be removed when the service is back.
In the longer term there a solution should be made that ensures this
dependency is available elsewhere.