fix(haproxy): re-order acl
lockpad uses `.well-known` as a public route, but the current
implementation would not route requests to it.

This change changes the order so that backends are chosen before well
known. The matrix backend remains after well-known because we want the
matrix well-known information to be used on both the base domain
`rubek.cloud` and `matrix.rubek.cloud`