fix(hyprland): make steam games work again
The issue involved using the hyprland flake directly without overriding
the nixpkgs version it uses. This means that the version of mesa used
was not the same one that my system was configured with and would result
in games being unable to launch from hyprland specifically.

I was able to figure this out thanks to https://github.com/NixOS/nixpkgs/issues/228586

It appears that steam still only launches properly when I start it from
a terminal, so there is further work to be done.