chore(hyprland): remove global keybind
This is now handled by the `global-keybind` module which provides a
compositor-agnostic method of doing the same thing. There are some
slight differences: this sends the key to any XWayland window and only
supports a single keybind but it is sufficient for now and allows plasma
on wayland to use the key.