fix: redeploy cluster
There was an issue performing updates on pyxis and ceylon. It is not
known why it did not succeed, but the decision was made to reprovision
them. When doing so, ceylon was downsized. Also, consul failed to
properly reform a cluster. It likely could have been fixed, but instead
bunky was also recycled and the `retry_join` blocks commented in favor
of manual joining.